add mountpoint=legacy
Been using this module in my config. It seems that in recent combinations of disko/systemd the latter can attempt to mount `/backups`, `/storage` and `/storage/save` while it was already been mounted by zfs, and fails dropping you into a shell (that sometimes you can just ctrl-d to resume from without issue, but still requires interacting with the server). 

The error you get is "Failed to mount local filesystems" on that target, with lines for each: "Failed to mount /storage", etc.

Thought to share so that a) you don't run into this issue if you ever reinstall b) get your feedback  as to whether this would pose any problems (so far none for me).